refactor: share dayname/monthname serde via spark-4.x shim; scope PR to those functions [skip ci]
The dayname/monthname conversion was identical across the spark-4.0/4.1/4.2
CometExprShim files. Move it into a single CometExprShim4x trait under the shared
spark-4.x source root (compiled for all 4.x profiles) and have each per-version
shim mix it in and delegate, removing the triplication.

Also drop the to_date/to_timestamp, make_timestamp, and current_date/now tests so
this PR is strictly the dayname/monthname feature; those cover rewrite-backed and
constant-folded functions whose documentation lives in the docs PR.
